log_analyzer/requirements.txt
slack_sdk
kafka-python
certifi
requests
urllib3